+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Please prefix each entry with one of:
 
 # Next Release
 
+- [fix] Log message when trace directory is not writable
+
 # v23.6.0
 - [feature] Support for uploading to artifactory with the Teamscale default artifact storage schema.
 - [feature] Display error window and add Event Log error entry when the Profiler.yml could not be loaded