+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Please prefix each entry with one of:
 
 
 # Next Release
-
+- [fix] Log message when trace directory is not writable
 - [feature] Added support to upload coverage to multiple Teamscale projects with different revisions. 
 
 # v23.6.0