+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Please prefix each entry with one of:
 
 
 # Next Release
+
+# v24.2.0
 - [fix] Log message when trace directory is not writable
 - [feature] Added support to upload coverage to multiple Teamscale projects with different revisions.